requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
Pillow>=10.0.0
openai>=1.0.0
requests>=2.31.0

# LangChain (optional but recommended)
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Web Search
duckduckgo-search>=4.0.0
google-search-results>=2.4.2

# Web Scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0

# PDF Processing
PyMuPDF>=1.23.0

# OCR for Images
pytesseract>=0.3.10

# Utilities
python-dateutil>=2.8.2